seperated chart components into parts
